Minor cleanup prior to next release - no logic changes
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@23343 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/alignments/Merge.java
@@ -224,7 +224,7 @@ public class Merge {
                     // with this specific query/read.
                 }
                 if (((wrote + skipped) % 1000000) == 0) {
-                    printStatus((int) wrote + skipped, wrote, skipped, skippedTooManyHits, skippedNotBestScore);
+                    printStatus(wrote + skipped, wrote, skipped, skippedTooManyHits, skippedNotBestScore);
                 }
 
             }
