+++ src/edu/cornell/med/icb/goby/reads/SequenceEncoder.java
@@ -26,7 +26,7 @@ package edu.cornell.med.icb.goby.reads;
 public final class SequenceEncoder {
    public void convertToByteBuffer(final byte[] sequence, final int referencePosition, final int readLength, final byte[] byteBuffer) {
         for (int i = 0; i < readLength; ++i) {
-            byteBuffer[i] = (byte) sequence[i + referencePosition];
+            byteBuffer[i] = sequence[i + referencePosition];
         }
     }
 }
